#56 moved filemgmt functions to helpers
