--- src/server/pkg/filemgmt/README.md
@@ -1,19 +1,0 @@
-# Files
-
-## Dependencies
-
-### Mediainfo
-
-Version v20.03 is required for the JSON output to work
-
-On MacOS
-```
-brew update && brew install mediainfo
-```
-
-On Ubuntu
-```
-sudo -s
-wget https://mediaarea.net/repo/deb/repo-mediaarea_1.0-12_all.deb && dpkg -i repo-mediaarea_1.0-12_all.deb && apt-get update
-apt-get install -y mediainfo
-```
